Split PartialEq into another file
