+++ src/free.ts
@@ -1,7 +1,7 @@
 import * as Option from "./option.js";
 import * as Result from "./result.js";
 
-import { Eq, PartialEq, eqSymbol } from "./type-class/eq.js";
+import { Eq, eqSymbol } from "./type-class/eq.js";
 import { Monad1, Monad2Monoid, kleisli } from "./type-class/monad.js";
 import type { Ord, PartialOrd } from "./type-class/ord.js";
 import { greater, less } from "./ordering.js";
@@ -9,6 +9,7 @@ import { greater, less } from "./ordering.js";
 import type { Applicative1 } from "./type-class/applicative.js";
 import type { Functor1 } from "./type-class/functor.js";
 import type { GetHktA1 } from "./hkt.js";
+import type { PartialEq } from "./type-class/partial-eq.js";
 import type { Traversable1 } from "./type-class/traversable.js";
 import type { Tuple } from "./tuple.js";
